Gracefully handle empty snippets
If snippet in Terraform code is optional, for example when ternary
operator is used, then provider is not able to handle at the moment and
crashes, trying to cast 'nil' interface into string.

This commit changes the logic to prevent that and makes that only non
'nil' snippets will be used.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>