feat: Add support to include files in configs
This commit implements support for Butane configs that include local
files as file or systemd-unit contents:

    variant: fcos
    version: "1.5.0"
    storage:
      files:
        - path: /some/path
          contents:
            local: ./some.file
    systemd:
      units:
        - name: some.service
          contents_local: ./some-unit.service

Since files are included relative to the directory specified by the
`FilesDir` argument, this only works if said argument is set.

As it stands today, Butane configs as outlined above will throw an error
message.

Signed-off-by: Manuel Hutter <manuel@hutter.io>